pytorch_implementation/requirements.txt
duplicate-finder
pytorch-lightning
opencv-python==4.5.*
numpy==1.21.*
matplotlib==3.5.*
pandas==1.3.*
pytest==6.2.*
pyyaml==6.0
torchvision
torchaudio
torch
albumentations==1.1.*
tensorboard
